checker: move healthchecks into separate package(s)
